+++ package.json
@@ -4,9 +4,12 @@
   "private": true,
   "dependencies": {
     "bootstrap": "^5.3.3",
+    "i18next": "^21.6.0",
+    "i18next-http-backend": "^3.0.1",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.7",
     "react-dom": "^18.3.1",
+    "react-i18next": "^11.11.0",
     "react-scripts": "^5.0.1",
     "react-scroll": "^1.9.0",
     "swiper": "^11.2.0"
