+++ src/components/about/styles.css
@@ -1,0 +1,9 @@
+
+.background-section {
+    width: 100%;
+    height: 100vh;
+    background-image: url(/public/pics/DSC_0583.jpg); /* Your image URL */
+    background-size: cover; /* Ensure the image covers the entire section */
+    background-position: center center; /* Centers the image */
+    background-repeat: no-repeat;
+}
