+++ src/components/home/home.js
@@ -1,0 +1,19 @@
+
+import './styles.css';
+
+const Home = () => {
+  return (
+    <div id="home" className="section">
+      <div className="background" ></div>
+      <div className="content">
+        <h3>Welcome to Vel Motifs</h3>
+        <p>Welcome to Vel Motifs, where creativity and tradition meet to create exquisite handcrafted designs. </p>
+        <p>What started as a hobby for a housewife is now blossoming into a future business, driven by a deep love for the ancient art of Aari work. </p>
+        <p>Our founder, Krithika Arumugam, has spent years perfecting her craft and is now ready to share this beautiful tradition with the world. </p>
+        <p>At Vel Motifs, we believe in the power of handmade artistry and the beauty that comes with each stitch. </p>
+        <p>We offer unique and custom-made Aari embroidery pieces that add elegance and charm to any occasion.</p>
+      </div>
+    </div>
+  );
+};
+  export default Home;
