+++ src/components/index.js
@@ -1,6 +1,6 @@
-export { default as Header } from './header';
-export { default as Footer } from './footer';
-export { default as Home } from './home';
-export { default as About } from './aboutme';
-export { default as Gallery } from './gallery';
-export { default as CollapsibleNavbar } from './collapsiblenavbar';
+export { default as Header } from './header/header';
+export { default as Footer } from './footer/footer';
+export { default as Home } from './home/home';
+export { default as About } from './about/aboutme';
+export { default as Gallery } from './gallery/gallery';
+export { default as CollapsibleNavbar } from './navbar/collapsiblenavbar';
