+++ src/components/navbar/styles.css
@@ -1,0 +1,7 @@
+.container, .sticky-top {
+    overflow: visible; /* Ensure the parent containers allow the sticky element */
+  }
+  
+  .text-right{
+    text-align: right;
+  }
